feature: Generate error codes from source
