+++ codegen/util.py
@@ -1,3 +1,5 @@
+import pathlib
+
 import pydantic
 
 
@@ -5,3 +7,8 @@ class BaseModel(
     pydantic.BaseModel,
     allow_mutation=False,
 ): ...
+
+
+def create_package(path: pathlib.Path) -> None:
+    path.mkdir(exist_ok=True)
+    (path / "__init__.py").touch(exist_ok=True)
