+++ java_tester/Dockerfile
@@ -5,9 +5,11 @@ WORKDIR /java_tester
 ADD src src
 ADD build.gradle build.gradle
 
-RUN gradle --no-daemon distTar && tar -xvf build/distributions/java_tester.tar
+RUN --mount=type=cache,target=/root/.gradle \
+    gradle --no-daemon distTar \
+ && tar -xvf build/distributions/java_tester.tar
 
 FROM eclipse-temurin:17-jre
 COPY --from=build /java_tester/java_tester /java_tester
 
-CMD /java_tester/bin/java_tester
+ENTRYPOINT ["/java_tester/bin/java_tester"]
