--- java_tester/Makefile
@@ -1,3 +1,0 @@
-.PHONY: docker_image
-docker_image:
-	docker build -t java_tester .
