+++ src/kio/serial/readers.py
@@ -10,7 +10,7 @@ from typing import TypeAlias
 from typing import TypeVar
 from uuid import UUID
 
-from kio.static.constants import ErrorCode
+from kio.schema.errors import ErrorCode
 from kio.static.constants import uuid_zero
 from kio.static.primitive import TZAware
 from kio.static.primitive import i8
