+++ src/kio/serial/writers.py
@@ -12,7 +12,7 @@ from typing import TypeAlias
 from typing import TypeVar
 from uuid import UUID
 
-from kio.static.constants import ErrorCode
+from kio.schema.errors import ErrorCode
 from kio.static.constants import uuid_zero
 from kio.static.primitive import i8
 from kio.static.primitive import i16
